+++ packages/form-render/src/form-items/tree-select/index.tsx
@@ -12,7 +12,6 @@ function useSelectOptions(): [Ref<SelectOptions>, (value: SelectOptions) => void
   return [selectOptions, updateSelectOptions]
 }
 export function renderTreeSelectItem<T=DataRecord>(options: RenderTreeSelectItemOptions) {
-  console.log('开始渲染SelectTree:')
   let selectInstance: ComponentPublicInstance
   let mounted = false
 
@@ -48,23 +47,21 @@ export function renderTreeSelectItem<T=DataRecord>(options: RenderTreeSelectItem
 
   const updateSelectOptionsFromCache = async () => {
     const value = cache.get(options.options)
-    console.log('当前缓存值:', value)
+
     switch (typeof value) {
       // 已经执行未返回
       case 'function':
-        console.log('检测到缓存正在请求中')
-        console.log('更新缓存')
+
         cache.set(options.options, updateSelectOptions)
         return
       // 已经执行已返回
       case 'object':
-        console.log('检测到缓存数据完成')
+
         updateSelectOptions(value)
         return
     }
     // 缓存执行
 
-    console.log('未检测到缓存开始进行请求')
     cache.set(options.options, updateSelectOptions)
     // 未执行
     runSelectOptionsFunction(
@@ -89,11 +86,9 @@ export function renderTreeSelectItem<T=DataRecord>(options: RenderTreeSelectItem
   switch (true) {
     case options.options instanceof Function:{
       if (options.cache !== false) {
-        console.log('缓存开启')
         updateSelectOptionsFromCache()
       }
       else {
-        console.log('缓存关闭')
         updateSelectOptionsFromData()
       }
       break