+++ packages/table-render/src/table-render/table-column-render.tsx
@@ -66,7 +66,7 @@ export function renderTableColumn<T>(options: TableColumnOptions<T>, pageMode: '
     dataIndex: options.index || options.key as string,
     title: options.title,
     width: options.width === 'auto' ? undefined : options.width,
-    align: options.align ?? 'center',
+    align: options.align ?? 'left',
     fixed: options.fixed,
     ellipsis: options.ellipsis ?? true,
     sortable: options.sortable